+++ context/app/static/js/pages/Home/HomeRevision.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import AssayTypeBarChartContainer from 'js/components/home-revision/AssayTypeBarChartContainer';
+
+function HomeRevision() {
+  return <AssayTypeBarChartContainer />;
+}
+
+export default HomeRevision;