+++ context/app/static/js/pages/Sample/Sample.jsx
@@ -14,7 +14,7 @@ import useEntityStore from 'js/stores/useEntityStore';
 import { AppContext } from 'js/components/Providers';
 import DetailContext from 'js/components/Detail/context';
 import { getSectionOrder } from 'js/components/Detail/utils';
-import useSearchHits from 'js/hooks/useSearchHits';
+import { useSearchHits } from 'js/hooks/useSearchData';
 import SampleSpecificDatasetsTable from 'js/components/Detail/SampleSpecificDatasetsTable';
 
 const entityStoreSelector = (state) => state.setAssayMetadata;
@@ -57,21 +57,13 @@ function SampleDetail(props) {
           ],
         },
       },
-      _source: [
-        'uuid',
-        'display_doi',
-        'mapped_data_types',
-        'descendant_counts',
-        'origin_sample.mapped_organ',
-        'status',
-        'last_modified_timestamp',
-      ],
+      _source: ['uuid', 'display_doi', 'mapped_data_types', 'status', 'descendant_counts', 'last_modified_timestamp'],
       size: 10000,
     }),
     [uuid],
   );
 
-  const { searchData: sampleSpecificDatasets } = useSearchHits(
+  const { searchHits: sampleSpecificDatasets } = useSearchHits(
     sampleSpecificDatasetsQuery,
     elasticsearchEndpoint,
     nexusToken,
@@ -84,7 +76,7 @@ function SampleDetail(props) {
     datasets: sampleSpecificDatasets.length > 0,
   };
   const sectionOrder = getSectionOrder(
-    ['summary', 'datasets', 'tissue', 'attribution', 'provenance', 'protocols', 'metadata'],
+    ['summary', 'datasets', 'tissue', 'provenance', 'protocols', 'metadata', 'attribution'],
     shouldDisplaySection,
   );
 
@@ -121,14 +113,14 @@ function SampleDetail(props) {
           mapped_organ={mapped_organ}
           hasRUI={hasRUI}
         />
+        <ProvSection uuid={uuid} assayMetadata={assayMetadata} />
+        {shouldDisplaySection.protocols && <Protocol protocol_url={protocol_url} />}
+        {shouldDisplaySection.metadata && <MetadataTable metadata={metadata} display_doi={display_doi} />}
         <Attribution
           group_name={group_name}
           created_by_user_displayname={created_by_user_displayname}
           created_by_user_email={created_by_user_email}
         />
-        <ProvSection uuid={uuid} assayMetadata={assayMetadata} />
-        {shouldDisplaySection.protocols && <Protocol protocol_url={protocol_url} />}
-        {shouldDisplaySection.metadata && <MetadataTable metadata={metadata} display_doi={display_doi} />}
       </DetailLayout>
     </DetailContext.Provider>
   );