+++ context/app/static/js/schema.org.js
@@ -1,16 +1,21 @@
 function getDatasetLD(entity) {
   // Based on https://developers.google.com/search/docs/data-types/dataset#guidelines
-  const donor = entity.donor.mapped_metadata;
-  const shortDonorString = `${donor.sex || '(Unknown sex)'}, ${donor.age_value || '(Unknown age)'} ${
-    donor.age_unit || '(Unknown age unit)'
-  } old`;
-  const heightString = `${donor.height_value || ''}${donor.height_unit || ''}`;
-  const weightString = `${donor.weight_value || ''}${donor.weight_unit || ''}`;
-  const longDonorString = `${heightString}, ${weightString}, ${donor.race || ''} ${shortDonorString}`;
   const assayOrganString = `${entity.mapped_data_types} of ${entity.origin_sample.mapped_organ}`;
-  const medicalHistory = donor.medicalHistory ? `${donor.medicalHistory.join(', ')}` : 'no medical history';
-  const name = `${assayOrganString} from ${shortDonorString}`;
-  const fallbackDescription = `${assayOrganString} from ${longDonorString} with ${medicalHistory}`;
+  let name = `${assayOrganString} from unknown donor`;
+  let fallbackDescription = name;
+
+  const donor = entity.donor?.mapped_metadata;
+  if (donor) {
+    const shortDonorString = `${donor.sex || '(Unknown sex)'}, ${donor.age_value || '(Unknown age)'} ${
+      donor.age_unit || '(Unknown age unit)'
+    } old`;
+    const heightString = `${donor.height_value || ''} ${donor.height_unit || ''}`;
+    const weightString = `${donor.weight_value || ''} ${donor.weight_unit || ''}`;
+    const longDonorString = `${heightString}, ${weightString}, ${donor.race || ''} ${shortDonorString}`;
+    const medicalHistory = donor.medicalHistory ? `${donor.medicalHistory.join(', ')}` : 'no medical history';
+    name = `${assayOrganString} from ${shortDonorString}`;
+    fallbackDescription = `${assayOrganString} from ${longDonorString} with ${medicalHistory}`;
+  }
 
   return {
     '@context': 'https://schema.org/',